Optionally limit on number of items in the cache.
You can now decide if older cache items should be removed by passing in
a new command line option `--keep-items`:

    package-cache --keep-items=5 install yarn

In this example, for any cache items found for the current version of
Yarn installed at the time of exectution, the most recent 5 will be
kept and others deleted.

Note that this only occurs when a new item is added to the cache and
that the `--noArchive` option is NOT present.

To help facilitate this feature, the `lodash` package has been replaced
with `underscore`.